fix: separate profile bites from home bites
