+++ libs/bite-tribe/home/data-access/src/lib/home-data-access.service.ts
@@ -1,4 +1,4 @@
-import { inject, Injectable, signal } from '@angular/core';
+import { inject, Injectable } from '@angular/core';
 import { BiteTribeStoreService } from 'bite-tribe/store';
 import { toSignal } from '@angular/core/rxjs-interop';
 import type { Bite, Like } from 'model';
@@ -16,7 +16,9 @@ export class HomeDataAccessService {
   sorting = toSignal(this.storeService.homeSorting$, {
     initialValue: 'distance',
   });
-  myBites = signal(<Bite[]>[]);
+  myBites = toSignal(this.storeService.sortedMyBites$, {
+    initialValue: [] as Bite[],
+  });
   myBitesSorting = toSignal(this.storeService.myBitesSorting$, {
     initialValue: 'distance',
   });
