+++ libs/bite-tribe/profile/data-access/src/lib/profile-data-access.service.ts
@@ -1,15 +1,8 @@
-import {
-  inject,
-  Injectable,
-  resource,
-  ResourceLoader,
-  signal,
-} from '@angular/core';
+import { inject, Injectable, resource, ResourceLoader } from '@angular/core';
 import { toSignal } from '@angular/core/rxjs-interop';
 import { BiteTribeStoreService } from 'bite-tribe/store';
 import type { Bite, Like, PublicUser } from 'model';
 import { FirebaseFirestore } from '@capacitor-firebase/firestore';
-import { of } from 'rxjs';
 
 @Injectable({ providedIn: 'root' })
 export class ProfileDataAccessService {
@@ -25,7 +18,9 @@ export class ProfileDataAccessService {
 
   myUser = toSignal(this.storeService.publicUser$);
 
-  myBites = signal(<Bite[]>[]);
+  myBites = toSignal(this.storeService.sortedMyBites$, {
+    initialValue: [] as Bite[],
+  });
 
   private bites = toSignal(this.storeService.bites$, {
     initialValue: [] as Bite[],
