+++ libs/bite-tribe/store/src/lib/filtering-and-sorting/filtering-and-sorting-slice.model.ts
@@ -1,11 +1,16 @@
 export type FilteringAndSortingSlice = {
   sorting?: {
     home?: string;
-    myBites?: string;
+    mybites?: string;
     bucketlists?: string;
   };
   filtering?: {
     home?: {
+      filters: string[];
+      distance?: number;
+      maxPrice?: number;
+    };
+    mybites?: {
       filters: string[];
       distance?: number;
       maxPrice?: number;
